+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Optimize: An empty date is output instead of `01/01/1970`
 - Add: Option to set the color for `todo_due` and `todo_completed`
+- Add: `file` and `file_size` as fields for output
 
 ## v1.4.0 (2021-03-30)